requirements.txt
django-lazysignup
django-bootstrap4
